Fix error in autosuggest from broken query response
If a search from autosuggest doesn't work for any reason, it will return
false, which breaks the JS since it's expecting specific properties on
the returned object. This ensures there will be a valid REST response
without signaling to the script that the request was successful, which
stops further processing but still provides some developer feedback in
the response.